+++ libraries/CurieImu/src/BMI160.h
@@ -573,9 +573,9 @@ class BMI160Class {
         void readAcceleration(short& x, short& y, short& z);
         void readRotation(short& x, short& y, short& z);
 
-        int readAccelerometer(int axis);
-        int readGyro(int axis);
-        int readTemperature();
+        short readAccelerometer(int axis);
+        short readGyro(int axis);
+        short readTemperature();
 
         boolean shockDetected(int axis, int direction);
         boolean motionDetected(int axis, int direction);
